+++ commons/src/commonTest/kotlin/com/vitorpamplona/amethyst/commons/call/PeerSessionManagerTest.kt
@@ -487,61 +487,3 @@ class PeerSessionManagerTest {
         assertFalse(bobAccepted, "bob should ignore alice's offer")
     }
 }
-
-/**
- * Fake PeerSession that records all operations for test assertions.
- * No real WebRTC involved.
- */
-class FakePeerSession(
-    private var signalingState: SignalingState = SignalingState.STABLE,
-) : PeerSession {
-    val addedCandidates = mutableListOf<IceCandidateData>()
-    var lastRemoteDescription: Pair<SdpType, String>? = null
-    var rolledBack = false
-    var disposed = false
-    var lastCreatedOffer: String? = null
-    var lastCreatedAnswer: String? = null
-
-    override fun getSignalingState(): SignalingState = signalingState
-
-    override fun setRemoteDescription(
-        type: SdpType,
-        sdp: String,
-    ) {
-        lastRemoteDescription = type to sdp
-        if (type == SdpType.ANSWER) {
-            signalingState = SignalingState.STABLE
-        } else if (type == SdpType.OFFER) {
-            signalingState = SignalingState.HAVE_REMOTE_OFFER
-        }
-    }
-
-    override fun addIceCandidate(candidate: IceCandidateData) {
-        addedCandidates.add(candidate)
-    }
-
-    override fun createOffer(onSdpCreated: (String) -> Unit) {
-        signalingState = SignalingState.HAVE_LOCAL_OFFER
-        val sdp = "fake-offer-sdp"
-        lastCreatedOffer = sdp
-        onSdpCreated(sdp)
-    }
-
-    override fun createAnswer(onSdpCreated: (String) -> Unit) {
-        val sdp = "fake-answer-sdp"
-        lastCreatedAnswer = sdp
-        signalingState = SignalingState.STABLE
-        onSdpCreated(sdp)
-    }
-
-    override fun rollback(onDone: () -> Unit) {
-        rolledBack = true
-        signalingState = SignalingState.STABLE
-        onDone()
-    }
-
-    override fun dispose() {
-        disposed = true
-        signalingState = SignalingState.CLOSED
-    }
-}
